requirements.txt
altair
numpy
pandas
pandas-datareader
pillow
plotly
scipy
scikit-learn>=0.18
streamlit
yfinance